Move IDial and IListen interfaces into core project
