--- src/HttpOverStream.Client/IDial.cs
@@ -1,12 +1,0 @@
-﻿using System.IO;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace HttpOverStream.Client
-{
-    public interface IDial
-    {
-        ValueTask<Stream> DialAsync(HttpRequestMessage request, CancellationToken cancellationToken);
-    }
-}
